refactor: move settings menu to a file
