+++ src/main.js
@@ -14,6 +14,7 @@ import xssFilters from "xss-filters";
 import './style/main.less';
 import { showRunInfo } from "./util/run-info";
 import { levelUp } from "./util/level-up";
+import { getSettingsMenu } from "./util/settings";
 
 export const version = "v0.4.15";
 
@@ -1006,6 +1007,8 @@ document.getElementById("runData").addEventListener("close", () => {
 	document.getElementById("runDataWeapons").innerHTML = "";
 })
 
+// Snapshot mode
+
 let wasMuted = false;
 function prepareSnapshot() {
 	if (!document.getElementById("snapshot-show-hud").checked) {
@@ -1096,56 +1099,7 @@ function restart() {
 	startGame(0);
 	document.getElementById("gameOver").close();
 }
-function getSettingsMenu() {
-	let elem = document.createElement("div");
-	elem.id = "settings";
-	editableSettings.forEach(setting => {
-		if (setting.type == "checkbox") {
-			let settingElem = document.createElement("input");
-			settingElem.type = "checkbox";
-			settingElem.setAttribute("for", setting.var);
-			settingElem.checked = settingsStore.get(setting.var, settingsStore.options.default[setting.var]);
-			settingElem.addEventListener("change", () => {
-				settings[setting.var] = settingElem.checked;
-				settingsStore.set(setting.var, settingElem.checked);
-			});
 
-			let label = document.createElement("label");
-			label.appendChild(document.createTextNode(setting.name));
-			label.setAttribute("for", setting.var);
-			label.addEventListener("click", () => settingElem.click());
-
-			elem.appendChild(settingElem);
-			elem.appendChild(label);
-
-		} else if (setting.type == "select") {
-			let label = document.createElement("label");
-			label.appendChild(document.createTextNode(setting.name));
-			label.setAttribute("for", setting.var);
-
-			let select = document.createElement("select");
-			select.setAttribute("for", setting.var);
-			setting.options.forEach((option, optionI) => {
-				let label = setting.labels[optionI];
-				let opt = document.createElement("option");
-				if (option == settings[setting.var]) opt.selected = true;
-				opt.appendChild(document.createTextNode(label));
-				opt.value = option;
-				select.appendChild(opt);
-			});
-			select.addEventListener("change", () => {
-				settings[setting.var] = select.value;
-				if (setting.var == "starDetail") updateStars();
-				settingsStore.set(setting.var, select.value);
-			});
-
-			elem.appendChild(label);
-			elem.appendChild(select);
-		}
-		elem.appendChild(document.createElement("br"));
-	});
-	return elem;
-}
 function getSettings() {
 	settings = settingsStore.getAll();
 }
